float64->32

+++ classifier_lib.py
@@ -103,7 +103,7 @@ def get_log_ratio(discriminator, input, time, class_labels):
         log_ratio = discriminator(input, timesteps=time, condition=class_labels)
         # prediction = torch.clip(logits, 1e-5, 1. - 1e-5)
         # log_ratio = torch.log(prediction / (1. - prediction))
-        return log_ratio
+        return log_ratio.to(torch.float32)
 
 class vpsde():
     def __init__(self):
